refactor(channels): centralize setup contracts
