+++ webui/src/components/settings/ToggleButton.tsx
@@ -1,0 +1,48 @@
+import { cn } from "@/lib/utils";
+
+export function ToggleButton({
+  checked,
+  disabled,
+  onChange,
+  ariaLabel,
+  label,
+}: {
+  checked: boolean;
+  disabled?: boolean;
+  onChange: (checked: boolean) => void;
+  ariaLabel?: string;
+  label: string;
+}) {
+  return (
+    <button
+      type="button"
+      role="switch"
+      aria-checked={checked}
+      aria-label={ariaLabel ?? label}
+      disabled={disabled}
+      onClick={() => {
+        if (!disabled) onChange(!checked);
+      }}
+      className={cn(
+        "relative inline-flex h-[22px] w-[38px] shrink-0 items-center rounded-full p-[2px]",
+        "transition-colors duration-200 ease-out focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
+        checked
+          ? "bg-[#2997FF] shadow-[inset_0_0_0_1px_rgba(0,0,0,0.035)]"
+          : "bg-muted shadow-[inset_0_0_0_1px_rgba(0,0,0,0.035)] hover:bg-muted/80",
+        disabled && "cursor-default opacity-60",
+        disabled && checked && "hover:bg-[#2997FF]",
+        disabled && !checked && "hover:bg-muted",
+      )}
+    >
+      <span
+        aria-hidden
+        className={cn(
+          "h-[18px] w-[18px] rounded-full bg-background shadow-[0_1px_2px_rgba(0,0,0,0.18),0_2px_7px_rgba(0,0,0,0.11)]",
+          "transition-transform duration-200 ease-out",
+          checked ? "translate-x-[16px]" : "translate-x-0",
+        )}
+      />
+      <span className="sr-only">{label}</span>
+    </button>
+  );
+}
